+++ src/main/java/lotto/Application.java
@@ -1,10 +1,10 @@
 package lotto;
 
-import lotto.service.LottoService;
+import lotto.controller.LottoController;
 
 public class Application {
     public static void main(String[] args) {
-        LottoService lottoService = new LottoService();
-        lottoService.run();
+        LottoController lottoController = new LottoController();
+        lottoController.run();
     }
 }
